refactor(parser): extract shared types into types.ts, consistent comment style
